+++ packages/go-gen/tests/cosmwasm_std__IbcQuery.go
@@ -15,11 +15,16 @@ type ChannelQuery struct {
 	// optional argument
 	PortID string `json:"port_id,omitempty"`
 }
+type FeeEnabledChannelQuery struct {
+	ChannelID string `json:"channel_id"`
+	PortID    string `json:"port_id,omitempty"`
+}
 
 // IBCQuery defines a query request from the contract into the chain.
 // This is the counterpart of [IbcQuery](https://github.com/CosmWasm/cosmwasm/blob/v0.14.0-beta1/packages/std/src/ibc.rs#L61-L83).
 type IBCQuery struct {
-	PortID       *PortIDQuery       `json:"port_id,omitempty"`
-	ListChannels *ListChannelsQuery `json:"list_channels,omitempty"`
-	Channel      *ChannelQuery      `json:"channel,omitempty"`
+	PortID            *PortIDQuery            `json:"port_id,omitempty"`
+	ListChannels      *ListChannelsQuery      `json:"list_channels,omitempty"`
+	Channel           *ChannelQuery           `json:"channel,omitempty"`
+	FeeEnabledChannel *FeeEnabledChannelQuery `json:"fee_enabled_channel,omitempty"`
 }